Implemented async_comprehension coroutine

+++ 0x02-python_async_comprehension/1-async_comprehension.py
@@ -2,20 +2,16 @@
 """
 Async Comprehension Generates a list from an async comprehension
 """
-import asyncio
 from typing import List
-from typing import AsyncGenerator
-
-
 async_generator = __import__('0-async_generator').async_generator
 
 
-async def async_comprehension() -> List[int]:
+async def async_comprehension() -> List[float]:
     """
     Coroutine to collect 10 random numbers using async comprehension
     over async_generator.
 
     Returns:
-        List[int]: List of 10 random numbers.
+        List[float]: List of 10 random numbers.
     """
     return [i async for i in async_generator()]
